move git.LsFiles and git.ListFiles to gitserver client
