+++ internal/codeintel/dependencies/live/git.go
@@ -30,7 +30,7 @@ func (s *gitService) GetCommits(ctx context.Context, repoCommits []api.RepoCommi
 }
 
 func (s *gitService) LsFiles(ctx context.Context, repo api.RepoName, commits api.CommitID, pathspecs ...gitserver.Pathspec) ([]string, error) {
-	return git.LsFiles(ctx, s.db, s.checker, repo, commits, pathspecs...)
+	return gitserver.NewClient(s.db).LsFiles(ctx, s.checker, repo, commits, pathspecs...)
 }
 
 func (s *gitService) Archive(ctx context.Context, repo api.RepoName, opts gitserver.ArchiveOptions) (io.ReadCloser, error) {
