+++ internal/gitserver/commands.go
@@ -16,12 +16,12 @@ import (
 	"sync"
 	"time"
 
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
-
 	"github.com/go-git/go-git/v5/plumbing/format/config"
 	"github.com/golang/groupcache/lru"
+	"github.com/grafana/regexp"
 	"github.com/opentracing/opentracing-go/log"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 
 	"github.com/sourcegraph/go-diff/diff"
 
@@ -919,4 +919,72 @@ func runRevParse(ctx context.Context, cmd GitCommand, spec string) (api.CommitID
 		return "", gitdomain.BadCommitError{Spec: spec, Commit: commit, Repo: cmd.Repo()}
 	}
 	return commit, nil
+}
+
+// LsFiles returns the output of `git ls-files`
+func (c *ClientImplementor) LsFiles(ctx context.Context, checker authz.SubRepoPermissionChecker, repo api.RepoName, commit api.CommitID, pathspecs ...Pathspec) ([]string, error) {
+	// TODO: mocks
+	//if Mocks.LsFiles != nil {
+	//	return Mocks.LsFiles(repo, commit)
+	//}
+	args := []string{
+		"ls-files",
+		"-z",
+		"--with-tree",
+		string(commit),
+	}
+
+	if len(pathspecs) > 0 {
+		args = append(args, "--")
+		for _, pathspec := range pathspecs {
+			args = append(args, string(pathspec))
+		}
+	}
+
+	cmd := c.GitCommand(repo, args...)
+	out, err := cmd.CombinedOutput(ctx)
+	if err != nil {
+		return nil, errors.WithMessage(err, fmt.Sprintf("git command %v failed (output: %q)", cmd.Args(), out))
+	}
+
+	files := strings.Split(string(out), "\x00")
+	// Drop trailing empty string
+	if len(files) > 0 && files[len(files)-1] == "" {
+		files = files[:len(files)-1]
+	}
+	return filterPaths(ctx, repo, checker, files)
+}
+
+// ListFiles returns a list of root-relative file paths matching the given
+// pattern in a particular commit of a repository.
+func (c *ClientImplementor) ListFiles(ctx context.Context, repo api.RepoName, commit api.CommitID, pattern *regexp.Regexp, checker authz.SubRepoPermissionChecker) (_ []string, err error) {
+	cmd := c.GitCommand(repo, "ls-tree", "--name-only", "-r", string(commit), "--")
+
+	out, err := cmd.CombinedOutput(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	var matching []string
+	for _, path := range strings.Split(string(out), "\n") {
+		if pattern.MatchString(path) {
+			matching = append(matching, path)
+		}
+	}
+
+	return filterPaths(ctx, repo, checker, matching)
+}
+
+// 🚨 SECURITY: All git methods that deal with file or path access need to have
+// sub-repo permissions applied
+func filterPaths(ctx context.Context, repo api.RepoName, checker authz.SubRepoPermissionChecker, paths []string) ([]string, error) {
+	if !authz.SubRepoEnabled(checker) {
+		return paths, nil
+	}
+	a := actor.FromContext(ctx)
+	filtered, err := authz.FilterActorPaths(ctx, checker, a, repo, paths)
+	if err != nil {
+		return nil, errors.Wrap(err, "filtering paths")
+	}
+	return filtered, nil
 }
